# decision-agent/agent-backend/agent-factory/cmd/openapi-docs/generated/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Agent Factory 智能体工厂 API 文档",
        "title": "Agent Factory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:30777",
    "basePath": "/api/agent-factory",
    "paths": {
...
        "/v3/agent-tpl/{id}/publish-info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取模板发布信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "获取模板发布信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "403": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新模板发布信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "模板"
                ],
                "summary": "更新模板发布信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "403": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer {token} - 需要有效的 OAuth 访问令牌",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}